+++ packages/expo-maps/ios/AppleMapsView.swift
@@ -13,7 +13,7 @@ class AppleMapsViewProps: ExpoSwiftUI.ViewProps {
   @Field var cameraPosition: CameraPosition
   @Field var uiSettings: MapUISettings = MapUISettings()
   @Field var properties: MapProperties = MapProperties()
-  @Field var userInterfaceStyle: UserInterfaceStyle = .automatic
+  @Field var colorScheme: MapColorScheme = .automatic
   let onMapClick = EventDispatcher()
   let onMarkerClick = EventDispatcher()
   let onPolylineClick = EventDispatcher()