+++ packages/expo-maps/ios/AppleMapsViewiOS17.swift
@@ -157,7 +157,9 @@ struct AppleMapsViewiOS17: View, AppleMapsViewProtocol {
           state.hasInitializedCamera = true
         }
       }
-      .background(UserInterfaceStyleApplier(style: props.userInterfaceStyle.toUIUserInterfaceStyle()))
+      .let(props.colorScheme.toColorScheme()) { view, colorScheme in
+        view.environment(\.colorScheme, colorScheme)
+      }
     }
   }
 }