+++ packages/expo-maps/ios/AppleMapsViewiOS18.swift
@@ -246,7 +246,9 @@ struct AppleMapsViewiOS18: View, AppleMapsViewProtocol {
           state.hasInitializedCamera = true
         }
       }
-      .background(UserInterfaceStyleApplier(style: props.userInterfaceStyle.toUIUserInterfaceStyle()))
+      .let(props.colorScheme.toColorScheme()) { view, colorScheme in
+        view.environment(\.colorScheme, colorScheme)
+      }
     }
   }