+++ packages/expo-maps/src/index.ts
@@ -42,8 +42,8 @@ export namespace AppleMaps {
   export const MapStyleElevation = AppleTypes.AppleMapsMapStyleElevation;
   export type MapStyleElevation = AppleTypes.AppleMapsMapStyleElevation;
 
-  export const UserInterfaceStyle = AppleTypes.AppleMapsUserInterfaceStyle;
-  export type UserInterfaceStyle = AppleTypes.AppleMapsUserInterfaceStyle;
+  export const MapColorScheme = AppleTypes.AppleMapsColorScheme;
+  export type MapColorScheme = AppleTypes.AppleMapsColorScheme;
 
   export type Marker = AppleTypes.AppleMapsMarker;
   export type MapUISettings = AppleTypes.AppleMapsUISettings;