+++ packages/expo-maps/CHANGELOG.md
@@ -6,7 +6,7 @@
 
 ### 🎉 New features
 
-- [iOS] Add `userInterfaceStyle` prop to control map appearance (light/dark mode) independently from the app's color scheme. ([#42199](https://github.com/expo/expo/pull/42199) by [@huextrat](https://github.com/huextrat))
+- [iOS] Add `colorScheme` prop to control map appearance (light/dark mode). ([#42199](https://github.com/expo/expo/pull/42199) by [@huextrat](https://github.com/huextrat))
 
 ### 🐛 Bug fixes